+++ core/actionrouter/action.go
@@ -7,7 +7,7 @@ import (
 
 	"opensvc.com/opensvc/core/client"
 	"opensvc.com/opensvc/core/clientcontext"
-	"opensvc.com/opensvc/core/entrypoints/monitor"
+	"opensvc.com/opensvc/core/monitor"
 	"opensvc.com/opensvc/core/path"
 	"opensvc.com/opensvc/util/xerrors"
 )